Refactored backend tips routes
